backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt
gunicorn>=20.1.0,<21.0.0
whitenoise>=6.5.0,<7.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=7.0.0,<8.0.0
pytest-django>=4.0.0,<5.0.0